+++ Grbl_Esp32/src/GCode.cpp
@@ -24,6 +24,12 @@
 
 #include "Grbl.h"
 
+// Allow iteration over CoordIndex values
+CoordIndex& operator++(CoordIndex& i) {
+    i = static_cast<CoordIndex>(static_cast<uint8_t>(i) + 1);
+    return i;
+}
+
 // NOTE: Max line number is defined by the g-code standard to be 99999. It seems to be an
 // arbitrary value, and some GUIs may require more. So we increased it based on a max safe
 // value when converting a float (7.2 digit precision)s to an integer.
