+++ Grbl_Esp32/src/GCode.h
@@ -232,6 +232,32 @@ enum GCParserFlags {
     GCParserLaserIsMotion  = bit(7),
 };
 
+// Various places in the code access saved coordinate system data
+// by a small integer index according to the values below.
+enum CoordIndex : uint8_t{
+    Begin = 0,
+    G54 = Begin,
+    G55,
+    G56,
+    G57,
+    G58,
+    G59,
+    // To support 9 work coordinate systems it would be necessary to define
+    // the following 3 and modify GCode.cpp to support G59.1, G59.2, G59.3
+    // G59_1,
+    // G59_2,
+    // G59_3,
+    NWCSystems,
+    G28 = NWCSystems,
+    G30,
+    // G92_2,
+    // G92_3,
+    End,
+};
+
+// Allow iteration over CoordIndex values
+CoordIndex& operator ++ (CoordIndex& i);
+
 // NOTE: When this struct is zeroed, the 0 values in the above types set the system defaults.
 typedef struct {
     Motion   motion;     // {G0,G1,G2,G3,G38.2,G80}
